services/herald-api/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
redis
psycopg2-binary
